+++ src/markwright/utils/blockOf.ts
@@ -1,0 +1,13 @@
+import * as React from 'react';
+
+export default function divOf(str: string, tag: string = 'div') {
+  return {
+    react(node, output, state) {
+      return React.createElement(tag, {
+        children: output(node.content),
+        className: str,
+        key: node.id || state.key
+      });
+    }
+  };
+}